requirements.txt
fastapi
uvicorn
pydantic
pyjwt
httpx
pytest
